pos updated

+++ Point Of Sale/live-class/resources/views/components/auth/verify-otp-form.blade.php
@@ -6,7 +6,7 @@
                     <h4>ENTER OTP CODE</h4>
                     <br/>
                     <label>4 Digit Code Here</label>
-                    <input id="code" placeholder="Code" class="form-control" type="text"/>
+                    <input id="otp" placeholder="Code" class="form-control" type="text"/>
                     <br/>
                     <button onclick="VerifyOtp()"  class="btn w-100 float-end btn-primary">Next</button>
                 </div>
@@ -14,25 +14,29 @@
         </div>
     </div>
 </div>
-
 <script>
    async function VerifyOtp() {
-
-        let code=document.getElementById('code').value;
-        if(code.length!==4){
-            errorToast("4 Digit Verification Code Required !");
+        let otp = document.getElementById('otp').value;
+        if(otp.length !==4){
+           errorToast('Invalid OTP')
         }
         else{
-            let res=await axios.post("/verify-otp",{
-                otp:code,
+            showLoader();
+            let res=await axios.post('/verify-otp', {
+                otp: otp,
                 email:sessionStorage.getItem('email')
             })
-            if(res.status===200){
+            hideLoader();
+
+            if(res.status===200 && res.data['status']==='success'){
+                successToast(res.data['message'])
                 sessionStorage.clear();
-                window.location.href="/resetPassword";
+                setTimeout(() => {
+                    window.location.href='/resetPassword'
+                }, 1000);
             }
             else{
-                errorToast("Something Went Wrong !")
+                errorToast(res.data['message'])
             }
         }
     }
